+++ Rocket.Unturned/Player/UnturnedPlayer.cs
@@ -469,8 +469,7 @@ namespace Rocket.Unturned.Player
             }
             set
             {
-                player.skills.channel.send("tellExperience", ESteamCall.SERVER, ESteamPacket.UPDATE_RELIABLE_BUFFER, value);
-                player.skills.channel.send("tellExperience", ESteamCall.OWNER, ESteamPacket.UPDATE_RELIABLE_BUFFER, value);
+                player.skills.ServerSetExperience(value);
             }
         }