+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes should be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Changed
+- Replaced `tellBroken` invocation with `serverSetLegsBroken`. 
+
 ## 4.9.3.7 - 2021-02-26
 
 ### Changed